feature: files reorganized
just updated the files structure and created files for new "persona" module
