+++ usuario/src/main/java/com/project/demo/repository/IUsuarioRepository.java
@@ -1,9 +1,11 @@
-package com.project.demo.repository.usuarioRepository;
+package com.project.demo.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.project.demo.model.usuarioModel.usuarioModel;
+import com.project.demo.model.usuarioModel;
+
+;
 
 @Repository
 public interface IUsuarioRepository extends CrudRepository<usuarioModel, Integer>{
